backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.2
openpyxl==3.1.2
python-multipart==0.0.6
pydantic==2.4.2
scikit-learn==1.3.2
nltk==3.8.1
numpy>=1.21.0
openai>=1.30.0  # GPT-4.1 시리즈 지원을 위해 최신 버전 사용
anthropic==0.7.0
psutil==5.9.6  # 메모리 모니터링을 위해 추가
sqlalchemy==2.0.23
alembic==1.13.1
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 🔥 소셜 로그인 관련 패키지
authlib==1.2.1          # OAuth 클라이언트 라이브러리
httpx==0.24.1           # 비동기 HTTP 클라이언트